Funcionalidad como se pidio

+++ adoption/src/main/java/com/FDMF/web/api/PetControllerApi.java
@@ -1,8 +1,12 @@
 package com.FDMF.web.api;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -17,11 +21,13 @@ public class PetControllerApi {
 	@Autowired
     private PetRepository petRepo;
 
+	//Buscar todos los pets
     @GetMapping
     public Iterable<Pet> findAllPets() {
         return petRepo.findAll();
     }
-
+    
+    //Buscar los pets por ID
     @GetMapping("/{id}")
     public ResponseEntity<Pet> petById(@PathVariable("id") Long id) {
         Optional<Pet> optPet = petRepo.findById(id);
@@ -31,27 +37,47 @@ public class PetControllerApi {
         return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
     }
 
-    @GetMapping("/search")
-    public Iterable<Pet> petsByName(@RequestParam("name") String name) {
-        return petRepo.findByName(name);
+    //Buscar los pets por Nombre
+    @GetMapping("/name/{name}")
+    public List<Pet> petsByName(@PathVariable("name") String name) {
+        return petRepo.findByNameContaining(name);
     }
 
+    //Crear un nuevo pet
     @PostMapping(consumes = "application/json")
     @ResponseStatus(HttpStatus.CREATED)
     public Pet createPet(@RequestBody Pet pet) {
         return petRepo.save(pet);
     }
 
+    
     @PutMapping
     @ResponseStatus(HttpStatus.OK)
     public Pet updatePet(@RequestBody Pet pet) {
         return petRepo.save(pet);
     }
-
+    
     @DeleteMapping("/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deletePet(@PathVariable("id") Long id) {
-        petRepo.deleteById(id);
+    public ResponseEntity<Void> deletePet(@PathVariable("id") Long id) {
+        Optional<Pet> optPet = petRepo.findById(id);
+        if (optPet.isPresent()) {
+            petRepo.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
-}
 
+    @GetMapping("/youngest")
+    public List<Pet> getYoungestPets() {
+        return petRepo.findTop20YoungestPets();
+    }
+
+    @GetMapping("/page")
+    public Page<Pet> getPetsByPage(
+            @RequestParam(defaultValue = "0") int page,
+            @RequestParam(defaultValue = "5") int size) {
+        Pageable pageable = PageRequest.of(page, size);
+        return petRepo.findAllPets(pageable);
+    }
+    
+}
